NavigationDrawer CocaPods
1- install CocoaPods via Terminal.
2- add Navigation Drawer pod to Pod file then install it.
3- make NavigationViewController then BaseViewController(ViewController) following with SlidingViewController by adding Action Segue[kind: Present Modally] from Bar Button item of ViewController to SlidingViewController.

ViewController: has a bar button contains line 3 horizontal image.
SlidingViewController: contains a colorful view with close button to left.

4- Then I moved to the coding part:
ViewController class: has two IBActions homeButtonPressed and edgePanGesture (if we swipe the screen to show menu) and Extend the UIViewControllerTransitioningDelegate .
SlidingView class: has variable of interactor that will be passed from the ViewController and actions like  handleGesture(sender: UIPanGestureRecognizer) to handle sipe from any direction and closeBtnPressed(_ sender: Any) which close the menu if we pressed on the closed button.

5- Finally, I hooked up the @IBAction to StoryBoard starting from Screen Edge Pan Gesture Recognize in ViewController to Pan Gesture Recognizer in SlidingViewController.